Use paging to read review requests
Suprisingly, even though a high limit is set for a call to get the
review requests, no more than 200 results are returned. Therefore
we need to use paging to retrieve the results, using the query
max-results just to limit the final number of results, instead of
passing it to the query directly.

Affects #44: Ensure that all calls are using paging