Primitive support for uploading new diffs
Allow the subclipse plugin to contribute a new action to the diff
part . This allows the ui plugin to delegate the diff generation
and project matching.

The extension point is incomplete, and very likely to change.

The wizard does not yet support posting the updated diff.

Affects #30: Upload new diffs to existing reviews with Subclipse